Test harness update (#392)
* Start issue branch

* Fix docker daemon start logic

* version bump

* Simplify syntax makefile

* clean files order

* Less aggressive log process clean in darwin

* Use function call simpler syntax

* Consolidate targets for dependency lists

* Consolidate makefile logic

* use correct variable name

* reuse variable

* Update doc

* Deprecate old log agg

* Update for dep update

* Increase timeout

* Fix import

* dependency updates

* version bump

+++ build/build_scripts/print_docker_logs.sh
@@ -12,8 +12,12 @@ done
 
 echo
 
-for p in $(echo "SELECT pid FROM log_pids;" | sqlite3 buildcache.sqlite3);
-do
-    kill -9 "$p"
-done;
+kernel_name=$(uname -s)
+if [[ "$kernel_name" == "Linux" ]];
+then
+    for p in $(echo "SELECT pid FROM log_pids;" | sqlite3 buildcache.sqlite3);
+    do
+        kill -9 "$p"
+    done;
+fi;
 echo 'DELETE FROM log_pids;' | sqlite3 buildcache.sqlite3;
